Crate rusoto_sdb

Amazon SimpleDB is a web service providing the core database functions of data indexing and querying in the cloud. By offloading the time and effort associated with building and operating a web-scale database, SimpleDB provides developers the freedom to focus on application development.
A traditional, clustered relational database requires a sizable upfront capital outlay, is complex to design, and often requires extensive and repetitive database administration. Amazon SimpleDB is dramatically simpler, requiring no schema, automatically indexing your data and providing a simple API for storage and access. This approach eliminates the administrative burden of data modeling, index maintenance, and performance tuning. Developers gain access to this functionality within Amazon's proven computing environment, are able to scale instantly, and pay only for what they use.

Visit http://aws.amazon.com/simpledb/ for more information.

If you're using the service, you're probably looking for SimpleDbClient and SimpleDb.

Structs

Attribute

BatchDeleteAttributesRequest

BatchPutAttributesRequest

CreateDomainRequest

DeletableItem

DeleteAttributesRequest

DeleteDomainRequest

DomainMetadataRequest

DomainMetadataResult

GetAttributesRequest

GetAttributesResult

Item

ListDomainsRequest

ListDomainsResult

PutAttributesRequest

ReplaceableAttribute

ReplaceableItem

SelectRequest

SelectResult

SimpleDbClient
A client for the Amazon SimpleDB API.

UpdateCondition
Specifies the conditions under which data should be updated. If an update condition is specified for a request, the data will only be updated if the condition is satisfied. For example, if an attribute with a specific name and value exists, or if a specific attribute doesn't exist.

Enums

BatchDeleteAttributesError
Errors returned by BatchDeleteAttributes

BatchPutAttributesError
Errors returned by BatchPutAttributes

CreateDomainError
Errors returned by CreateDomain

DeleteAttributesError
Errors returned by DeleteAttributes

DeleteDomainError
Errors returned by DeleteDomain

DomainMetadataError
Errors returned by DomainMetadata

GetAttributesError
Errors returned by GetAttributes

ListDomainsError
Errors returned by ListDomains

PutAttributesError
Errors returned by PutAttributes

SelectError
Errors returned by Select

Traits

SimpleDb
Trait representing the capabilities of the Amazon SimpleDB API. Amazon SimpleDB clients implement this trait.