Struct rusoto_sdb::SimpleDbClient

pub struct SimpleDbClient { /* fields omitted */ }

A client for the Amazon SimpleDB API.

Methods

impl SimpleDbClient

pub fn new(region: Region) -> SimpleDbClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SimpleDbClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,